[241110] refactor: 검증 코드 분리 - 문자열 공백 확인 코드 분리 - 파일 사이즈 및 확장자 검증 코드 분리
