+++ 02-main_camp/Day43/src/app/_components/boards/Boards-write/hook.tsx
@@ -15,6 +15,7 @@ import CONSTANTS_ALERT_MESSAGE from "@/commons/constants/alert";
 import { Address } from "react-daum-postcode";
 import useModal from "@/commons/ui/modal/hook";
 import { NavigationPaths, useNavigate } from "@/commons/navigate";
+import { checkValidationImage, checkWithSpace } from "@/utils/validation";
 
 export const useBoardWrite = (isEdit?: boolean, data?: FetchBoardQuery) => {
   const navigate = useNavigate();
@@ -203,23 +204,6 @@ export const useBoardWrite = (isEdit?: boolean, data?: FetchBoardQuery) => {
     });
   };
 
-  const checkValidationImage = (file: File) => {
-    if (file.size > 5 * 1024 * 1024) {
-      alert("파일 용량이 너무 큽니다.(제한: 5MB)");
-      return false;
-    }
-
-    if (
-      !file.type.includes("jpeg") &&
-      !file.type.includes("png") &&
-      !file.type.includes("svg")
-    ) {
-      alert("jpeg 또는 png 파일만 업로드 가능합니다.");
-      return false;
-    }
-    return true;
-  };
-
   const onClickImage = (event: MouseEvent<HTMLDivElement>) => {
     // 어떤 요소를 눌렀는지 알아야한다.
     const id = Number(event.currentTarget.id);
@@ -239,15 +223,6 @@ export const useBoardWrite = (isEdit?: boolean, data?: FetchBoardQuery) => {
         images: newImages,
       };
     });
-  };
-
-  const checkWithSpace = (input: string) => {
-    const hasSpace = /\s/.test(input);
-    const isOnlySpace = input.trim().length === 0;
-
-    if (hasSpace) return true;
-    if (isOnlySpace) return true;
-    return false;
   };
 
   // 새로 고침을 해도 데이터를 불러오도록 구현
